edn: use BTree{Map, Set} to actually use a Set type
This had to be done to use a internal Set on the Edn::Set
variant.

Also the Edn::Map type started using a BTreeMap instead of the
original HashMap because of missing implementation of Hash.

The Edn::Double type also had to migrate to ordered-float crate
since it implements Ord and Hash.

Co-authored-by: Julia Boeira <julia.boeira@nubank.com.br>